fix issues found during the workshop + add workflow

+++ 2019/DW331343-L Dynamo for Software Developers/S2 - C# Fundamentals/Exercise 4/DynamoDev-S2-E4-Complete/DynamoDev-S2-E4-Complete/Utils/GeometryUtils.cs
@@ -7,7 +7,7 @@ namespace DynamoDev.Utils
 {
     public static class GeometryUtils
     {
-        public const int DEFAULT_NUMBER_OF_POINTS = 10;
+        public const double DEFAULT_NUMBER_OF_POINTS = 10.0;
 
         /// <summary>
         /// Generate a grid of equally spaced points on a given surface, flattened to a simple list.
@@ -15,14 +15,14 @@ namespace DynamoDev.Utils
         /// <param name="surface">The surface to generate the grid on.</param>
         /// <param name="numberOfPoints">The number of points in both X and Y coordinates, default to 10.</param>
         /// <returns>A flat list of points that cover the surface area.</returns>
-        internal static List<Point> GeneratePointGridForSurface(Surface surface, int numberOfPoints = DEFAULT_NUMBER_OF_POINTS)
+        internal static List<Point> GeneratePointGridForSurface(Surface surface, double numberOfPoints = DEFAULT_NUMBER_OF_POINTS)
         {
             if (surface is null)
                 throw new ArgumentNullException(nameof(surface));
 
             var listOfPoints = new List<Point>();
 
-            double step = 1 / numberOfPoints;
+            double step = 1.0 / numberOfPoints;
             for (double i = 0; i < 1; i += step)
             {
                 for (double j = 0; j < 1; j += step)
